Implement tool to generate CircleCI configuration
This commit contains the first version of a tool to generate CircleCI
configurations to test Apache Cassandra with dtest. The tool modifies
config-1_2.yaml in .circleci to run dtest from a fork and to repeat a
dtest with and/or without vnodes. It generates config.yaml from the
modifications to run with medium resources as it is minimal for dtest
execution. Dtest configuration is done for each given Cassandra branch
and vnodes/novnodes combination. It is achieved by creating a separate
branch for each case, committing with specific message and pushing
each branch to origin.

The tool also provides an option to remove branches by including
specific string.

It is the first prototype version. The code needs polishing. UI needs
improvements. It does certain assumptions, but doesn't check if they
are satisfied. These assumptions also put limitations on the tool
generality.

The tool is expected to evolved and external contributions might be
helpful.