Use option -e in the config generate tool
This commit modifies the tool to generate CircleCI configs and commits
to use option -e of .circleci/generate.sh. It assumes that all
branches, which will be tests, include the new version of generate,
i.e., the patch for CASSANDRA-16989 is merged in the corresponding
Cassandra branches.

This commit also adds option --remote to provide git remote for
Cassandra branches to test. If the option is not provided, then
local branches are used and commits are done directly there without
creating new branches.